[FIX] S3 코드 개선
